Extending MIS class for data mining - instructions
I added a comment explaining how to extend this application's information (from MIS class) into a meaningful way for data mining. 

In particular, the MIS class computes all statistics of the simulation, including a decision making algorithm that determines if a simulation is acceptable or not to use, given the exact price of the derivative. This decision can be mapped into {0,1}, 0 if we reject the simulated price, 1 otherwise, and along with the other members (information) of the class we can use getters to pull everything out into a csv file or a database. 

I believe it was a necessary addition for those who are interested in data analytics.